Chatbot/requirements.txt
fastapi
pydantic
langchain
langchain-core
langchain-community
langchain-huggingface
python-dotenv
uvicorn
bs4
requests
pandas
numpy
ast
faiss-cpu
sentence-transformers
hf-xet
pypdf
speechrecognition
yt-dlp
pydub
reportlab
Pillow
torch
transformers